vSphere: reorganization
Also disabled the untested sections.
