+++ builtin/providers/vsphere/consts.go
@@ -1,4 +1,4 @@
-package dvs
+package vsphere
 
 import "regexp"
