+++ builtin/providers/vsphere/off/f_maphostswitch.go
