+++ builtin/providers/vsphere/resource_vsphere_folder.go
@@ -1,4 +1,4 @@
-package folder
+package vsphere
 
 import (
 	"fmt"
@@ -6,7 +6,6 @@ import (
 	"path"
 	"strings"
 
-	"github.com/hashicorp/terraform/builtin/providers/vsphere/helpers"
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/vmware/govmomi"
 	"github.com/vmware/govmomi/find"
@@ -121,7 +120,7 @@ func resourceVSphereFolderRead(d *schema.ResourceData, meta interface{}) error {
 	log.Printf("[DEBUG] reading folder: %#v", d)
 	client := meta.(*govmomi.Client)
 
-	dc, err := helpers.GetDatacenter(client, d.Get("datacenter").(string))
+	dc, err := getDatacenter(client, d.Get("datacenter").(string))
 	if err != nil {
 		return err
 	}
@@ -167,7 +166,7 @@ func resourceVSphereFolderDelete(d *schema.ResourceData, meta interface{}) error
 }
 
 func deleteFolder(client *govmomi.Client, f *folder) error {
-	dc, err := helpers.GetDatacenter(client, f.datacenter)
+	dc, err := getDatacenter(client, f.datacenter)
 	if err != nil {
 		return err
 	}
