+++ builtin/providers/vsphere/resource_vsphere_folder_test.go
@@ -1,11 +1,10 @@
-package folder
+package vsphere
 
 import (
 	"fmt"
 	"os"
 	"testing"
 
-	"github.com/hashicorp/terraform/builtin/providers/vsphere/helpers"
 	"github.com/hashicorp/terraform/helper/resource"
 	"github.com/hashicorp/terraform/terraform"
 	"github.com/vmware/govmomi"
@@ -23,8 +22,8 @@ func TestAccVSphereFolder_basic(t *testing.T) {
 	path := "tf_test_basic"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { helpers.TestAccPreCheck(t) },
-		Providers:    helpers.TestAccProviders,
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckVSphereFolderDestroy,
 		Steps: []resource.TestStep{
 			resource.TestStep{
@@ -55,8 +54,8 @@ func TestAccVSphereFolder_nested(t *testing.T) {
 	path := "tf_test_nested/tf_test_folder"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:     func() { helpers.TestAccPreCheck(t) },
-		Providers:    helpers.TestAccProviders,
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckVSphereFolderDestroy,
 		Steps: []resource.TestStep{
 			resource.TestStep{
@@ -88,8 +87,8 @@ func TestAccVSphereFolder_dontDeleteExisting(t *testing.T) {
 	path := existingPath + "/tf_nested/tf_test"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:  func() { helpers.TestAccPreCheck(t) },
-		Providers: helpers.TestAccProviders,
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
 		CheckDestroy: resource.ComposeTestCheckFunc(
 			assertVSphereFolderExists(datacenter, existingPath),
 			removeVSphereFolder(datacenter, existingPath, ""),
@@ -118,7 +117,7 @@ func TestAccVSphereFolder_dontDeleteExisting(t *testing.T) {
 }
 
 func testAccCheckVSphereFolderDestroy(s *terraform.State) error {
-	client := helpers.TestAccProvider.Meta().(*govmomi.Client)
+	client := testAccProvider.Meta().(*govmomi.Client)
 	finder := find.NewFinder(client.Client, true)
 
 	for _, rs := range s.RootModule().Resources {
@@ -160,7 +159,7 @@ func testAccCheckVSphereFolderExists(n string, f *folder) resource.TestCheckFunc
 			return fmt.Errorf("No ID is set")
 		}
 
-		client := helpers.TestAccProvider.Meta().(*govmomi.Client)
+		client := testAccProvider.Meta().(*govmomi.Client)
 		finder := find.NewFinder(client.Client, true)
 
 		dc, err := finder.Datacenter(context.TODO(), rs.Primary.Attributes["datacenter"])
@@ -194,7 +193,7 @@ func testAccCheckVSphereFolderExistingPathExists(n string, f *folder) resource.T
 			return fmt.Errorf("No ID is set")
 		}
 
-		client := helpers.TestAccProvider.Meta().(*govmomi.Client)
+		client := testAccProvider.Meta().(*govmomi.Client)
 		finder := find.NewFinder(client.Client, true)
 
 		dc, err := finder.Datacenter(context.TODO(), rs.Primary.Attributes["datacenter"])
@@ -220,7 +219,7 @@ func testAccCheckVSphereFolderExistingPathExists(n string, f *folder) resource.T
 func assertVSphereFolderExists(datacenter string, folder_name string) resource.TestCheckFunc {
 
 	return func(s *terraform.State) error {
-		client := helpers.TestAccProvider.Meta().(*govmomi.Client)
+		client := testAccProvider.Meta().(*govmomi.Client)
 		folder, err := object.NewSearchIndex(client.Client).FindByInventoryPath(
 			context.TODO(), fmt.Sprintf("%v/vm/%v", datacenter, folder_name))
 		if err != nil {
@@ -235,7 +234,7 @@ func assertVSphereFolderExists(datacenter string, folder_name string) resource.T
 
 func createVSphereFolder(datacenter string, folder_name string) error {
 
-	client := helpers.TestAccProvider.Meta().(*govmomi.Client)
+	client := testAccProvider.Meta().(*govmomi.Client)
 
 	f := folder{path: folder_name, datacenter: datacenter}
 
@@ -260,7 +259,7 @@ func removeVSphereFolder(datacenter string, folder_name string, existing_path st
 
 	return func(s *terraform.State) error {
 
-		client := helpers.TestAccProvider.Meta().(*govmomi.Client)
+		client := testAccProvider.Meta().(*govmomi.Client)
 		// finder := find.NewFinder(client.Client, true)
 
 		folder, _ := object.NewSearchIndex(client.Client).FindByInventoryPath(
