+++ builtin/providers/vsphere/resource_vsphere_virtual_disk_test.go
@@ -1,4 +1,4 @@
-package virtual_disk
+package vsphere
 
 import (
 	"fmt"
