+++ builtin/providers/vsphere/resource_vsphere_virtual_machine_test.go
@@ -1,4 +1,4 @@
-package virtual_machine
+package vsphere
 
 import (
 	"fmt"
